Update with latest generated files.

+++ src/main/java/yokwe/majuro/type/INTEGER.java
@@ -8,15 +8,15 @@ public final class INTEGER extends MemoryData16 {
     public static final Class<?> SELF = java.lang.invoke.MethodHandles.lookup().lookupClass();
     public static final String   NAME = SELF.getSimpleName();
     
-    public static final int  WORD_SIZE =               1;
-    public static final int  BIT_SIZE  =              16;
-                                                         
-    public static final long MIN_VALUE = Short.MIN_VALUE;
-    public static final long MAX_VALUE = Short.MAX_VALUE;
+    public static final int WORD_SIZE =               1;
+    public static final int BIT_SIZE  =              16;
+                                                        
+    public static final int MIN_VALUE = Short.MIN_VALUE;
+    public static final int MAX_VALUE = Short.MAX_VALUE;
     
     private static final ContextSubrange context = new ContextSubrange(NAME, MIN_VALUE, MAX_VALUE);
     
-    public static final void checkValue(long value) {
+    public static final void checkValue(int value) {
         if (Debug.ENABLE_CHECK_VALUE) context.check(value);
     }
